[Refactor] 카드 모달 구조 분리
